% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_login_page.R
\name{create_login_page}
\alias{create_login_page}
\title{Create login page}
\usage{
create_login_page(
  app_path = ".",
  yaml_file = "./firebase.yaml",
  overwrite = FALSE
)
}
\arguments{
\item{app_path}{Path of your shiny app file.}

\item{overwrite}{Logical. Overwrite files if they exist. Defaults to FALSE.}

\item{yaml_path}{Path to your firebase.yaml file.}
}
\value{
Invisible TRUE.
}
\description{
Create a default login page.
}
