% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{firebaseLogoutButton}
\alias{firebaseLogoutButton}
\title{Firebase logout button}
\usage{
firebaseLogoutButton(label = "Log out", ..., id = "._firebaselogout_")
}
\arguments{
\item{label}{The contents of the button or link–usually a text label,
but you could also use any other HTML, like an image.}

\item{...}{Named attributes to be applied to the button or link.}

\item{id}{An ID for the button. If you only have one logout button in your
app, you do not need to explicitly provide an ID. If you have more than one
logout button, you need to provide a unique ID to each button. When you
create a button with a non-default ID, you must create an observer that
listens to a click on this button and logs out of the app with a call to
logout.}
}
\description{
Firebase logout button
}
