% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyAppFirebase}
\alias{shinyAppFirebase}
\title{Create a Firebase Shiny app object}
\usage{
shinyAppFirebase(ui, server, ...)
}
\arguments{
\item{ui}{The ui object of your app.}

\item{server}{Ther server function of your app.}

\item{...}{Additional parameters passed to shiny::shinyApp function.}
}
\value{
An object that represents the app. Printing the object or passing
it to runApp() will run the app.
}
\description{
Create a Firebase Shiny app object
}
