% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_firebase.R
\name{create_shiny_firebase}
\alias{create_shiny_firebase}
\title{Create shiny firebase app}
\usage{
create_shiny_firebase(path, overwrite = FALSE)
}
\arguments{
\item{path}{Path where the app file will be created.}

\item{overwrite}{Logical. Overwrite files if they exist. Defaults to FALSE.}
}
\description{
Create shiny firebase app
}
